Diplom project add events jobs and users

+++ app/src/main/java/ru/netology/nework/util/FormatNumber.kt
@@ -1,12 +1,12 @@
 package ru.netology.nework.util
 
 object FormatNumber {
-    fun formatNumber(count: Long): String {
+    fun formatNumber(count: Int): String {
         return when {
             count < 1000 -> count.toString()
             count < 10000 -> {
                 val hundreds = (count % 1000) / 100
-                if (hundreds == 0.toLong()) {
+                if (hundreds == 0) {
                     "${count / 1000}K"
                 } else {
                     "${count / 1000}.${hundreds}K"
@@ -16,7 +16,7 @@ object FormatNumber {
             count < 1000000 -> "${count / 1000}K"
             else -> {
                 val hundredThousands = (count % 1000000) / 100000
-                if (hundredThousands == 0.toLong()) {
+                if (hundredThousands == 0) {
                     "${count / 1000000}M"
                 } else {
                     "${count / 1000000}.${hundredThousands}M"
